Added DAO to change Reps, Weight and Sets
This DAO will allow the change of Reps and Weights or both in the database, the set number cannot be logically written into this database for an exercise, it must be coded in the front end. Do NOT use the set id as the set number it is not there for this purpose.